# docs/swagger.yaml
basePath: /api/v1
definitions:
  model.SureSureBank:
    properties:
      account_no:
        type: string
      account_type:
        type: string
      bank_code:
        type: string
      created_date:
        type: string
      id:
        type: integer
      is_active:
        type: integer
      name_en:
        type: string
      name_th:
        type: string
      prompt_pay_type:
        type: string
      updated_date:
        type: string
      user_id:
        type: integer
    type: object
  model.SureSureLog:
    properties:
      action:
        type: string
      created_date:
        type: string
      data_request:
        type: string
      id:
        type: integer
      method_name:
        type: string
      name_th:
        type: string
      updated_date:
        type: string
      user_id:
        type: integer
    type: object
  model.SureSureOrderPackage:
    properties:
      created_date:
        type: string
      id:
        type: integer
      package_id:
        type: integer
      price:
        type: number
      ref_no:
        type: string
      status:
        type: string
      updated_date:
        type: string
      user_id:
        type: integer
    type: object
...
  util.APIResponse:
    properties:
      data: {}
      message:
        type: string
      status_code:
        type: integer
      timestamp:
        type: string
    type: object
info:
  contact: {}
  description: Public API for SureSure services
  title: SureSure Public API
  version: "1.0"
paths:
  /bank2/create:
    post:
      consumes:
      - application/json
      parameters:
      - description: Bank payload
        in: body
        name: bank
        required: true
        schema:
          $ref: '#/definitions/model.SureSureBank'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.APIResponse'
      summary: Create bank
      tags:
      - Bank
  /bank2/delete/{id}:
    delete:
      parameters:
      - description: Bank ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.APIResponse'
      summary: Delete bank by ID
      tags:
      - Bank
  /bank2/get:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.APIResponse'
      summary: List banks
      tags:
      - Bank
  /bank2/get/{id}:
    get:
      parameters:
      - description: Bank ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.APIResponse'
      summary: Get bank by ID
      tags:
      - Bank
  /bank2/update:
    put:
      consumes:
      - application/json
      parameters:
      - description: Bank payload
        in: body
        name: bank
        required: true
        schema:
          $ref: '#/definitions/model.SureSureBank'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.APIResponse'
      summary: Update bank
      tags:
      - Bank
...
  /log/create:
    post:
      consumes:
      - application/json
      parameters:
      - description: Log payload
        in: body
        name: log
        required: true
        schema:
          $ref: '#/definitions/model.SureSureLog'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.APIResponse'
      summary: Create log
      tags:
      - Log
  /log/delete/{id}:
    delete:
      parameters:
      - description: Log ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.APIResponse'
      summary: Delete log by ID
      tags:
      - Log
  /log/get:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.APIResponse'
      summary: List logs
      tags:
      - Log
  /log/get/{id}:
    get:
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.APIResponse'
      summary: Get log by user ID
      tags:
      - Log
  /log/update:
    put:
      consumes:
      - application/json
      parameters:
      - description: Log payload
        in: body
        name: log
        required: true
        schema:
          $ref: '#/definitions/model.SureSureLog'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.APIResponse'
      summary: Update log
      tags:
      - Log
...
  /order-package/create:
    post:
      consumes:
      - application/json
      parameters:
      - description: Order package payload
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/model.SureSureOrderPackage'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.APIResponse'
      summary: Create order package
      tags:
      - OrderPackage
  /order-package/delete/{id}:
    delete:
      parameters:
      - description: Order package ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.APIResponse'
      summary: Delete order package by ID
      tags:
      - OrderPackage
  /order-package/get:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.APIResponse'
      summary: List order packages
      tags:
      - OrderPackage
  /order-package/get/{id}:
    get:
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.APIResponse'
      summary: Get order package by user ID
      tags:
      - OrderPackage
  /order-package/get/refno/{RefNo}:
    get:
      parameters:
      - description: Reference No
        in: path
        name: RefNo
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.APIResponse'
      summary: Get order package by RefNo
      tags:
      - OrderPackage
  /order-package/update:
    put:
      consumes:
      - application/json
      parameters:
      - description: Order package payload
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/model.SureSureOrderPackage'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.APIResponse'
      summary: Update order package
      tags:
      - OrderPackage
...
schemes:
- http
- https
swagger: "2.0"